backend/requirements.txt
fastapi
uvicorn
openai
vecs
python-dotenv
psycopg2-binary
